requirements.txt
python-dotenv
requests
langchain-openai
beautifulsoup4